refactor: 💡 Rename products to product-pages
