+++ apps/webstore/src/app/app.component.spec.ts
@@ -5,7 +5,7 @@ import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { AppComponent } from './app.component';
 import { HomeComponent } from './pages/home/home.component';
 import { PageNotFoundComponent } from './pages/issues/page-not-found/page-not-found.component';
-import { FamilyTreeComponent } from './pages/catalog/products/family-tree/family-tree.component';
+import { FamilyTreeComponent } from './pages/catalog/product-pages/family-tree/family-tree.component';
 import { FooterComponent } from './shared/components/footer/footer.component';
 import { CookiePromptModalComponent } from './shared/components/modals/cookie-prompt-modal/cookie-prompt-modal.component';
 import { NavbarComponent } from './shared/components/navbar/navbar.component';
