+++ apps/webstore/src/app/pages/catalog/product-pages/family-tree/family-tree.component.mobile.scss
