+++ apps/webstore/src/app/pages/catalog/product-pages/quotable/quotable.component.mobile.scss
